backend/requirements.txt
# Core Framework
Django==6.0.2
asgiref==3.11.1
tzdata==2025.3

# Database (PostgreSQL - Modern Driver)
psycopg[binary]==3.1.18
sqlparse==0.5.5

# REST API & Security
djangorestframework==3.16.1
djangorestframework-simplejwt==5.3.1
django-cors-headers==4.3.1
django-filter==25.2

# Environment & Static Files
python-dotenv==1.0.1
whitenoise==6.12.0

# Asynchronous Tasks
celery==5.3.6
redis==5.0.3

# Document Generation
weasyprint>=68.0

# Production Server
gunicorn==21.2.0
uvicorn[standard]>=0.29.0

# API Documentation
drf-spectacular==0.27.1

# Error Monitoring
sentry-sdk==1.40.6

# --- Environment & Configuration ---
django-environ==0.12.0

# --- Email & Communication ---
django-anymail[resend]==13.0

# --- Logging & Observability ---
python-json-logger==3.2.1

pydantic==2.12.5
email-validator>=2.1.1

requests==2.31.0